secures delete item button

+++ backend/src/resolvers/Mutation.js
@@ -43,8 +43,15 @@ const Mutations = {
     const where = { id: args.id };
     const item = await ctx.db.query.item(
       { where }, 
-      `{ id title }`
+      `{ id title user {id}}`
     );
+    const ownsItem = item.user.id == ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => (
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    ));
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to delete items that aren\'t yours.');
+    } 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
